prjct/requirements.txt
JayDeBeApi==1.2.3
Werkzeug==1.0.1
apache-airflow-providers-amazon==5.0.0
apache-airflow-providers-celery==2.1.0
apache-airflow-providers-jdbc==3.2.0
apache-airflow-providers-odbc==3.1.1
apache-airflow-providers-postgres==2.4.0
apache-airflow-providers-redis==2.0.1
apache-airflow-providers-common-sql==1.1.0
boto3==1.18.65
boto==2.49.0
botocore==1.21.65
numpy==1.20.3
pandas==1.3.5
pyodbc==4.0.34
pyarrow==5.0.0
pyOpenSSL==20.0.1
psycopg2-binary==2.9.2
redis==3.5.3
sshtunnel==0.4.0
virtualenv==20.10.0
cloudpathlib[s3,gs,azure]
apache-airflow-providers-yandex==3.1.0
yandexcloud==0.183.0
xlrd==2.0.1
s3fs
geopandas
GeoAlchemy2


